Move addWithCheckBody to BaseProcessorUnit
